pyproject.toml
[tool.poetry]
name = "vyaguta-chatbot"
version = "0.1.0"
description = "Vyaguta Onboarding Assistant Chatbot (Python, LangChain, RAG, OpenAI)"
authors = ["Leapfrog AI COE"]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.2.0"
openai = "^1.30.1"
tiktoken = "^0.7.0"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
